% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{Time that the customer arrives}

\item{ServiceTimes}{Time that the customer spent on getting the service done}

\item{NumServers}{Number of server}
}
\value{
Customers' transition time
}
\description{
Feed customers though a multiserver queue system to determine each customer's transition time.
}
\examples{
Multiserver (Arrivals = 1, ServiceTimes = 1, NumServers = 2)
}
